+++ src/firmware_new/boards/frankie_v1/crc.c
@@ -1,0 +1,78 @@
+/**
+ ******************************************************************************
+ * File Name          : CRC.c
+ * Description        : This file provides code for the configuration
+ *                      of the CRC instances.
+ ******************************************************************************
+ * @attention
+ *
+ * <h2><center>&copy; Copyright (c) 2020 STMicroelectronics.
+ * All rights reserved.</center></h2>
+ *
+ * This software component is licensed by ST under Ultimate Liberty license
+ * SLA0044, the "License"; You may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at:
+ *                             www.st.com/SLA0044
+ *
+ ******************************************************************************
+ */
+
+/* Includes ------------------------------------------------------------------*/
+#include "crc.h"
+
+/* USER CODE BEGIN 0 */
+
+/* USER CODE END 0 */
+
+CRC_HandleTypeDef hcrc;
+
+/* CRC init function */
+void MX_CRC_Init(void)
+{
+    hcrc.Instance                     = CRC;
+    hcrc.Init.DefaultPolynomialUse    = DEFAULT_POLYNOMIAL_ENABLE;
+    hcrc.Init.DefaultInitValueUse     = DEFAULT_INIT_VALUE_ENABLE;
+    hcrc.Init.InputDataInversionMode  = CRC_INPUTDATA_INVERSION_NONE;
+    hcrc.Init.OutputDataInversionMode = CRC_OUTPUTDATA_INVERSION_DISABLE;
+    hcrc.InputDataFormat              = CRC_INPUTDATA_FORMAT_BYTES;
+    if (HAL_CRC_Init(&hcrc) != HAL_OK)
+    {
+        Error_Handler();
+    }
+}
+
+void HAL_CRC_MspInit(CRC_HandleTypeDef* crcHandle)
+{
+    if (crcHandle->Instance == CRC)
+    {
+        /* USER CODE BEGIN CRC_MspInit 0 */
+
+        /* USER CODE END CRC_MspInit 0 */
+        /* CRC clock enable */
+        __HAL_RCC_CRC_CLK_ENABLE();
+        /* USER CODE BEGIN CRC_MspInit 1 */
+
+        /* USER CODE END CRC_MspInit 1 */
+    }
+}
+
+void HAL_CRC_MspDeInit(CRC_HandleTypeDef* crcHandle)
+{
+    if (crcHandle->Instance == CRC)
+    {
+        /* USER CODE BEGIN CRC_MspDeInit 0 */
+
+        /* USER CODE END CRC_MspDeInit 0 */
+        /* Peripheral clock disable */
+        __HAL_RCC_CRC_CLK_DISABLE();
+        /* USER CODE BEGIN CRC_MspDeInit 1 */
+
+        /* USER CODE END CRC_MspDeInit 1 */
+    }
+}
+
+/* USER CODE BEGIN 1 */
+
+/* USER CODE END 1 */
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
