+++ src/firmware_new/boards/frankie_v1/main.h
@@ -51,8 +51,6 @@ extern "C"
 
     /* USER CODE END EM */
 
-    void HAL_TIM_MspPostInit(TIM_HandleTypeDef *htim);
-
     /* Exported functions prototypes ---------------------------------------------*/
     void Error_Handler(void);
