+++ src/firmware_new/boards/frankie_v1/stm32h7xx_hal_conf.h
@@ -163,54 +163,53 @@ extern "C"
 #define USE_RTOS 0U
 #define USE_SD_TRANSCEIVER 0U /*!< use uSD Transceiver */
 
-#define USE_HAL_ADC_REGISTER_CALLBACKS 0U    /* ADC register callback disabled     */
-#define USE_HAL_CEC_REGISTER_CALLBACKS 0U    /* CEC register callback disabled     */
-#define USE_HAL_COMP_REGISTER_CALLBACKS 0U   /* COMP register callback disabled    */
-#define USE_HAL_CRYP_REGISTER_CALLBACKS 0U   /* CRYP register callback disabled    */
-#define USE_HAL_DAC_REGISTER_CALLBACKS 0U    /* DAC register callback disabled     */
-#define USE_HAL_DCMI_REGISTER_CALLBACKS 0U   /* DCMI register callback disabled    */
-#define USE_HAL_DFSDM_REGISTER_CALLBACKS 0U  /* DFSDM register callback disabled   */
-#define USE_HAL_DMA2D_REGISTER_CALLBACKS 0U  /* DMA2D register callback disabled   */
-#define USE_HAL_DSI_REGISTER_CALLBACKS 0U    /* DSI register callback disabled     */
-#define USE_HAL_DTS_REGISTER_CALLBACKS 0U    /* DTS register callback disabled     */
-#define USE_HAL_ETH_REGISTER_CALLBACKS 0U    /* ETH register callback disabled     */
-#define USE_HAL_FDCAN_REGISTER_CALLBACKS 0U  /* FDCAN register callback disabled   */
-#define USE_HAL_NAND_REGISTER_CALLBACKS 0U   /* NAND register callback disabled    */
-#define USE_HAL_NOR_REGISTER_CALLBACKS 0U    /* NOR register callback disabled     */
-#define USE_HAL_SDRAM_REGISTER_CALLBACKS 0U  /* SDRAM register callback disabled   */
-#define USE_HAL_SRAM_REGISTER_CALLBACKS 0U   /* SRAM register callback disabled    */
-#define USE_HAL_HASH_REGISTER_CALLBACKS 0U   /* HASH register callback disabled    */
-#define USE_HAL_HCD_REGISTER_CALLBACKS 0U    /* HCD register callback disabled     */
-#define USE_HAL_GFXMMU_REGISTER_CALLBACKS 0U /* GFXMMU register callback disabled  */
-#define USE_HAL_HRTIM_REGISTER_CALLBACKS 0U  /* HRTIM register callback disabled   */
-#define USE_HAL_I2C_REGISTER_CALLBACKS 0U    /* I2C register callback disabled     */
-#define USE_HAL_I2S_REGISTER_CALLBACKS 0U    /* I2S register callback disabled     */
-#define USE_HAL_IRDA_REGISTER_CALLBACKS 0U   /* IRDA register callback disabled    */
-#define USE_HAL_JPEG_REGISTER_CALLBACKS 0U   /* JPEG register callback disabled    */
-#define USE_HAL_LPTIM_REGISTER_CALLBACKS 0U  /* LPTIM register callback disabled   */
-#define USE_HAL_LTDC_REGISTER_CALLBACKS 0U   /* LTDC register callback disabled    */
-#define USE_HAL_MDIOS_REGISTER_CALLBACKS 0U  /* MDIO register callback disabled    */
-#define USE_HAL_MMC_REGISTER_CALLBACKS 0U    /* MMC register callback disabled     */
-#define USE_HAL_OPAMP_REGISTER_CALLBACKS 0U  /* MDIO register callback disabled    */
-#define USE_HAL_OSPI_REGISTER_CALLBACKS 0U   /* OSPI register callback disabled    */
-#define USE_HAL_OTFDEC_REGISTER_CALLBACKS 0U /* OTFDEC register callback disabled  */
-#define USE_HAL_PCD_REGISTER_CALLBACKS 0U    /* PCD register callback disabled     */
-#define USE_HAL_QSPI_REGISTER_CALLBACKS 0U   /* QSPI register callback disabled    */
-#define USE_HAL_RNG_REGISTER_CALLBACKS 0U    /* RNG register callback disabled     */
-#define USE_HAL_RTC_REGISTER_CALLBACKS 0U    /* RTC register callback disabled     */
-#define USE_HAL_SAI_REGISTER_CALLBACKS 0U    /* SAI register callback disabled     */
-#define USE_HAL_SD_REGISTER_CALLBACKS 0U     /* SD register callback disabled      */
-#define USE_HAL_SMARTCARD_REGISTER_CALLBACKS                                             \
-    0U                                        /* SMARTCARD register callback disabled    \
-                                               */
-#define USE_HAL_SPDIFRX_REGISTER_CALLBACKS 0U /* SPDIFRX register callback disabled */
-#define USE_HAL_SMBUS_REGISTER_CALLBACKS 0U   /* SMBUS register callback disabled   */
-#define USE_HAL_SPI_REGISTER_CALLBACKS 0U     /* SPI register callback disabled     */
-#define USE_HAL_SWPMI_REGISTER_CALLBACKS 0U   /* SWPMI register callback disabled   */
-#define USE_HAL_TIM_REGISTER_CALLBACKS 0U     /* TIM register callback disabled     */
-#define USE_HAL_UART_REGISTER_CALLBACKS 0U    /* UART register callback disabled    */
-#define USE_HAL_USART_REGISTER_CALLBACKS 0U   /* USART register callback disabled   */
-#define USE_HAL_WWDG_REGISTER_CALLBACKS 0U    /* WWDG register callback disabled    */
+#define USE_HAL_ADC_REGISTER_CALLBACKS 0U       /* ADC register callback disabled     */
+#define USE_HAL_CEC_REGISTER_CALLBACKS 0U       /* CEC register callback disabled     */
+#define USE_HAL_COMP_REGISTER_CALLBACKS 0U      /* COMP register callback disabled    */
+#define USE_HAL_CRYP_REGISTER_CALLBACKS 0U      /* CRYP register callback disabled    */
+#define USE_HAL_DAC_REGISTER_CALLBACKS 0U       /* DAC register callback disabled     */
+#define USE_HAL_DCMI_REGISTER_CALLBACKS 0U      /* DCMI register callback disabled    */
+#define USE_HAL_DFSDM_REGISTER_CALLBACKS 0U     /* DFSDM register callback disabled   */
+#define USE_HAL_DMA2D_REGISTER_CALLBACKS 0U     /* DMA2D register callback disabled   */
+#define USE_HAL_DSI_REGISTER_CALLBACKS 0U       /* DSI register callback disabled     */
+#define USE_HAL_DTS_REGISTER_CALLBACKS 0U       /* DTS register callback disabled     */
+#define USE_HAL_ETH_REGISTER_CALLBACKS 0U       /* ETH register callback disabled     */
+#define USE_HAL_FDCAN_REGISTER_CALLBACKS 0U     /* FDCAN register callback disabled   */
+#define USE_HAL_NAND_REGISTER_CALLBACKS 0U      /* NAND register callback disabled    */
+#define USE_HAL_NOR_REGISTER_CALLBACKS 0U       /* NOR register callback disabled     */
+#define USE_HAL_SDRAM_REGISTER_CALLBACKS 0U     /* SDRAM register callback disabled   */
+#define USE_HAL_SRAM_REGISTER_CALLBACKS 0U      /* SRAM register callback disabled    */
+#define USE_HAL_HASH_REGISTER_CALLBACKS 0U      /* HASH register callback disabled    */
+#define USE_HAL_HCD_REGISTER_CALLBACKS 0U       /* HCD register callback disabled     */
+#define USE_HAL_GFXMMU_REGISTER_CALLBACKS 0U    /* GFXMMU register callback disabled  */
+#define USE_HAL_HRTIM_REGISTER_CALLBACKS 0U     /* HRTIM register callback disabled   */
+#define USE_HAL_I2C_REGISTER_CALLBACKS 0U       /* I2C register callback disabled     */
+#define USE_HAL_I2S_REGISTER_CALLBACKS 0U       /* I2S register callback disabled     */
+#define USE_HAL_IRDA_REGISTER_CALLBACKS 0U      /* IRDA register callback disabled    */
+#define USE_HAL_JPEG_REGISTER_CALLBACKS 0U      /* JPEG register callback disabled    */
+#define USE_HAL_LPTIM_REGISTER_CALLBACKS 0U     /* LPTIM register callback disabled   */
+#define USE_HAL_LTDC_REGISTER_CALLBACKS 0U      /* LTDC register callback disabled    */
+#define USE_HAL_MDIOS_REGISTER_CALLBACKS 0U     /* MDIO register callback disabled    */
+#define USE_HAL_MMC_REGISTER_CALLBACKS 0U       /* MMC register callback disabled     */
+#define USE_HAL_OPAMP_REGISTER_CALLBACKS 0U     /* MDIO register callback disabled    */
+#define USE_HAL_OSPI_REGISTER_CALLBACKS 0U      /* OSPI register callback disabled    */
+#define USE_HAL_OTFDEC_REGISTER_CALLBACKS 0U    /* OTFDEC register callback disabled  */
+#define USE_HAL_PCD_REGISTER_CALLBACKS 0U       /* PCD register callback disabled     */
+#define USE_HAL_QSPI_REGISTER_CALLBACKS 0U      /* QSPI register callback disabled    */
+#define USE_HAL_RNG_REGISTER_CALLBACKS 0U       /* RNG register callback disabled     */
+#define USE_HAL_RTC_REGISTER_CALLBACKS 0U       /* RTC register callback disabled     */
+#define USE_HAL_SAI_REGISTER_CALLBACKS 0U       /* SAI register callback disabled     */
+#define USE_HAL_SD_REGISTER_CALLBACKS 0U        /* SD register callback disabled      */
+#define USE_HAL_SMARTCARD_REGISTER_CALLBACKS 0U /* SMARTCARD register callback disabled  \
+                                                 */
+#define USE_HAL_SPDIFRX_REGISTER_CALLBACKS 0U   /* SPDIFRX register callback disabled */
+#define USE_HAL_SMBUS_REGISTER_CALLBACKS 0U     /* SMBUS register callback disabled   */
+#define USE_HAL_SPI_REGISTER_CALLBACKS 0U       /* SPI register callback disabled     */
+#define USE_HAL_SWPMI_REGISTER_CALLBACKS 0U     /* SWPMI register callback disabled   */
+#define USE_HAL_TIM_REGISTER_CALLBACKS 0U       /* TIM register callback disabled     */
+#define USE_HAL_UART_REGISTER_CALLBACKS 0U      /* UART register callback disabled    */
+#define USE_HAL_USART_REGISTER_CALLBACKS 0U     /* USART register callback disabled   */
+#define USE_HAL_WWDG_REGISTER_CALLBACKS 0U      /* WWDG register callback disabled    */
 
 /* ########################### Ethernet Configuration ######################### */
 #define ETH_TX_DESC_CNT 4 /* number of Ethernet Tx DMA descriptors */
