+++ src/firmware_new/boards/frankie_v1/stm32h7xx_hal_msp.c
@@ -58,8 +58,6 @@
 /* USER CODE BEGIN 0 */
 
 /* USER CODE END 0 */
-
-void HAL_TIM_MspPostInit(TIM_HandleTypeDef* htim);
 /**
  * Initializes the Global MSP.
  */
@@ -78,266 +76,6 @@ void HAL_MspInit(void)
     /* USER CODE BEGIN MspInit 1 */
 
     /* USER CODE END MspInit 1 */
-}
-
-/**
- * @brief CRC MSP Initialization
- * This function configures the hardware resources used in this example
- * @param hcrc: CRC handle pointer
- * @retval None
- */
-void HAL_CRC_MspInit(CRC_HandleTypeDef* hcrc)
-{
-    if (hcrc->Instance == CRC)
-    {
-        /* USER CODE BEGIN CRC_MspInit 0 */
-
-        /* USER CODE END CRC_MspInit 0 */
-        /* Peripheral clock enable */
-        __HAL_RCC_CRC_CLK_ENABLE();
-        /* USER CODE BEGIN CRC_MspInit 1 */
-
-        /* USER CODE END CRC_MspInit 1 */
-    }
-}
-
-/**
- * @brief CRC MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param hcrc: CRC handle pointer
- * @retval None
- */
-void HAL_CRC_MspDeInit(CRC_HandleTypeDef* hcrc)
-{
-    if (hcrc->Instance == CRC)
-    {
-        /* USER CODE BEGIN CRC_MspDeInit 0 */
-
-        /* USER CODE END CRC_MspDeInit 0 */
-        /* Peripheral clock disable */
-        __HAL_RCC_CRC_CLK_DISABLE();
-        /* USER CODE BEGIN CRC_MspDeInit 1 */
-
-        /* USER CODE END CRC_MspDeInit 1 */
-    }
-}
-
-/**
- * @brief TIM_PWM MSP Initialization
- * This function configures the hardware resources used in this example
- * @param htim_pwm: TIM_PWM handle pointer
- * @retval None
- */
-void HAL_TIM_PWM_MspInit(TIM_HandleTypeDef* htim_pwm)
-{
-    if (htim_pwm->Instance == TIM4)
-    {
-        /* USER CODE BEGIN TIM4_MspInit 0 */
-
-        /* USER CODE END TIM4_MspInit 0 */
-        /* Peripheral clock enable */
-        __HAL_RCC_TIM4_CLK_ENABLE();
-        /* USER CODE BEGIN TIM4_MspInit 1 */
-
-        /* USER CODE END TIM4_MspInit 1 */
-    }
-}
-
-void HAL_TIM_MspPostInit(TIM_HandleTypeDef* htim)
-{
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (htim->Instance == TIM4)
-    {
-        /* USER CODE BEGIN TIM4_MspPostInit 0 */
-
-        /* USER CODE END TIM4_MspPostInit 0 */
-
-        __HAL_RCC_GPIOD_CLK_ENABLE();
-        __HAL_RCC_GPIOB_CLK_ENABLE();
-        /**TIM4 GPIO Configuration
-        PD13     ------> TIM4_CH2
-        PD14     ------> TIM4_CH3
-        PD15     ------> TIM4_CH4
-        PB6     ------> TIM4_CH1
-        */
-        GPIO_InitStruct.Pin = wheel_motor_back_left_pwm_Pin |
-                              wheel_motor_back_right_pwm_Pin |
-                              wheel_motor_front_right_pwm_Pin;
-        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull      = GPIO_NOPULL;
-        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF2_TIM4;
-        HAL_GPIO_Init(GPIOD, &GPIO_InitStruct);
-
-        GPIO_InitStruct.Pin       = wheel_motor_front_left_pwm_Pin;
-        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull      = GPIO_NOPULL;
-        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF2_TIM4;
-        HAL_GPIO_Init(wheel_motor_front_left_pwm_GPIO_Port, &GPIO_InitStruct);
-
-        /* USER CODE BEGIN TIM4_MspPostInit 1 */
-
-        /* USER CODE END TIM4_MspPostInit 1 */
-    }
-}
-/**
- * @brief TIM_PWM MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param htim_pwm: TIM_PWM handle pointer
- * @retval None
- */
-void HAL_TIM_PWM_MspDeInit(TIM_HandleTypeDef* htim_pwm)
-{
-    if (htim_pwm->Instance == TIM4)
-    {
-        /* USER CODE BEGIN TIM4_MspDeInit 0 */
-
-        /* USER CODE END TIM4_MspDeInit 0 */
-        /* Peripheral clock disable */
-        __HAL_RCC_TIM4_CLK_DISABLE();
-        /* USER CODE BEGIN TIM4_MspDeInit 1 */
-
-        /* USER CODE END TIM4_MspDeInit 1 */
-    }
-}
-
-/**
- * @brief UART MSP Initialization
- * This function configures the hardware resources used in this example
- * @param huart: UART handle pointer
- * @retval None
- */
-void HAL_UART_MspInit(UART_HandleTypeDef* huart)
-{
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (huart->Instance == USART3)
-    {
-        /* USER CODE BEGIN USART3_MspInit 0 */
-
-        /* USER CODE END USART3_MspInit 0 */
-        /* Peripheral clock enable */
-        __HAL_RCC_USART3_CLK_ENABLE();
-
-        __HAL_RCC_GPIOD_CLK_ENABLE();
-        /**USART3 GPIO Configuration
-        PD8     ------> USART3_TX
-        PD9     ------> USART3_RX
-        */
-        GPIO_InitStruct.Pin       = STLK_RX_Pin | STLK_TX_Pin;
-        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull      = GPIO_NOPULL;
-        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF7_USART3;
-        HAL_GPIO_Init(GPIOD, &GPIO_InitStruct);
-
-        /* USER CODE BEGIN USART3_MspInit 1 */
-
-        /* USER CODE END USART3_MspInit 1 */
-    }
-}
-
-/**
- * @brief UART MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param huart: UART handle pointer
- * @retval None
- */
-void HAL_UART_MspDeInit(UART_HandleTypeDef* huart)
-{
-    if (huart->Instance == USART3)
-    {
-        /* USER CODE BEGIN USART3_MspDeInit 0 */
-
-        /* USER CODE END USART3_MspDeInit 0 */
-        /* Peripheral clock disable */
-        __HAL_RCC_USART3_CLK_DISABLE();
-
-        /**USART3 GPIO Configuration
-        PD8     ------> USART3_TX
-        PD9     ------> USART3_RX
-        */
-        HAL_GPIO_DeInit(GPIOD, STLK_RX_Pin | STLK_TX_Pin);
-
-        /* USER CODE BEGIN USART3_MspDeInit 1 */
-
-        /* USER CODE END USART3_MspDeInit 1 */
-    }
-}
-
-/**
- * @brief PCD MSP Initialization
- * This function configures the hardware resources used in this example
- * @param hpcd: PCD handle pointer
- * @retval None
- */
-void HAL_PCD_MspInit(PCD_HandleTypeDef* hpcd)
-{
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (hpcd->Instance == USB_OTG_FS)
-    {
-        /* USER CODE BEGIN USB_OTG_FS_MspInit 0 */
-
-        /* USER CODE END USB_OTG_FS_MspInit 0 */
-
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-        /**USB_OTG_FS GPIO Configuration
-        PA8     ------> USB_OTG_FS_SOF
-        PA9     ------> USB_OTG_FS_VBUS
-        PA10     ------> USB_OTG_FS_ID
-        PA11     ------> USB_OTG_FS_DM
-        PA12     ------> USB_OTG_FS_DP
-        */
-        GPIO_InitStruct.Pin       = USB_SOF_Pin | USB_ID_Pin | USB_DM_Pin | USB_DP_Pin;
-        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull      = GPIO_NOPULL;
-        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF10_OTG1_FS;
-        HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-        GPIO_InitStruct.Pin  = USB_VBUS_Pin;
-        GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        HAL_GPIO_Init(USB_VBUS_GPIO_Port, &GPIO_InitStruct);
-
-        /* Peripheral clock enable */
-        __HAL_RCC_USB_OTG_FS_CLK_ENABLE();
-        /* USER CODE BEGIN USB_OTG_FS_MspInit 1 */
-
-        /* USER CODE END USB_OTG_FS_MspInit 1 */
-    }
-}
-
-/**
- * @brief PCD MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param hpcd: PCD handle pointer
- * @retval None
- */
-void HAL_PCD_MspDeInit(PCD_HandleTypeDef* hpcd)
-{
-    if (hpcd->Instance == USB_OTG_FS)
-    {
-        /* USER CODE BEGIN USB_OTG_FS_MspDeInit 0 */
-
-        /* USER CODE END USB_OTG_FS_MspDeInit 0 */
-        /* Peripheral clock disable */
-        __HAL_RCC_USB_OTG_FS_CLK_DISABLE();
-
-        /**USB_OTG_FS GPIO Configuration
-        PA8     ------> USB_OTG_FS_SOF
-        PA9     ------> USB_OTG_FS_VBUS
-        PA10     ------> USB_OTG_FS_ID
-        PA11     ------> USB_OTG_FS_DM
-        PA12     ------> USB_OTG_FS_DP
-        */
-        HAL_GPIO_DeInit(
-            GPIOA, USB_SOF_Pin | USB_VBUS_Pin | USB_ID_Pin | USB_DM_Pin | USB_DP_Pin);
-
-        /* USER CODE BEGIN USB_OTG_FS_MspDeInit 1 */
-
-        /* USER CODE END USB_OTG_FS_MspDeInit 1 */
-    }
 }
 
 /* USER CODE BEGIN 1 */
