+++ src/firmware_new/boards/frankie_v1/tim.c
@@ -1,0 +1,146 @@
+/**
+ ******************************************************************************
+ * File Name          : TIM.c
+ * Description        : This file provides code for the configuration
+ *                      of the TIM instances.
+ ******************************************************************************
+ * @attention
+ *
+ * <h2><center>&copy; Copyright (c) 2020 STMicroelectronics.
+ * All rights reserved.</center></h2>
+ *
+ * This software component is licensed by ST under Ultimate Liberty license
+ * SLA0044, the "License"; You may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at:
+ *                             www.st.com/SLA0044
+ *
+ ******************************************************************************
+ */
+
+/* Includes ------------------------------------------------------------------*/
+#include "tim.h"
+
+/* USER CODE BEGIN 0 */
+
+/* USER CODE END 0 */
+
+TIM_HandleTypeDef htim4;
+
+/* TIM4 init function */
+void MX_TIM4_Init(void)
+{
+    TIM_MasterConfigTypeDef sMasterConfig = {0};
+    TIM_OC_InitTypeDef sConfigOC          = {0};
+
+    htim4.Instance               = TIM4;
+    htim4.Init.Prescaler         = 9 - 1;
+    htim4.Init.CounterMode       = TIM_COUNTERMODE_UP;
+    htim4.Init.Period            = 400 - 1;
+    htim4.Init.ClockDivision     = TIM_CLOCKDIVISION_DIV1;
+    htim4.Init.AutoReloadPreload = TIM_AUTORELOAD_PRELOAD_DISABLE;
+    if (HAL_TIM_PWM_Init(&htim4) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    sMasterConfig.MasterOutputTrigger = TIM_TRGO_RESET;
+    sMasterConfig.MasterSlaveMode     = TIM_MASTERSLAVEMODE_DISABLE;
+    if (HAL_TIMEx_MasterConfigSynchronization(&htim4, &sMasterConfig) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    sConfigOC.OCMode     = TIM_OCMODE_PWM1;
+    sConfigOC.Pulse      = 0;
+    sConfigOC.OCPolarity = TIM_OCPOLARITY_HIGH;
+    sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
+    if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_1) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_2) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_3) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_4) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    HAL_TIM_MspPostInit(&htim4);
+}
+
+void HAL_TIM_PWM_MspInit(TIM_HandleTypeDef* tim_pwmHandle)
+{
+    if (tim_pwmHandle->Instance == TIM4)
+    {
+        /* USER CODE BEGIN TIM4_MspInit 0 */
+
+        /* USER CODE END TIM4_MspInit 0 */
+        /* TIM4 clock enable */
+        __HAL_RCC_TIM4_CLK_ENABLE();
+        /* USER CODE BEGIN TIM4_MspInit 1 */
+
+        /* USER CODE END TIM4_MspInit 1 */
+    }
+}
+void HAL_TIM_MspPostInit(TIM_HandleTypeDef* timHandle)
+{
+    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    if (timHandle->Instance == TIM4)
+    {
+        /* USER CODE BEGIN TIM4_MspPostInit 0 */
+
+        /* USER CODE END TIM4_MspPostInit 0 */
+
+        __HAL_RCC_GPIOD_CLK_ENABLE();
+        __HAL_RCC_GPIOB_CLK_ENABLE();
+        /**TIM4 GPIO Configuration
+        PD13     ------> TIM4_CH2
+        PD14     ------> TIM4_CH3
+        PD15     ------> TIM4_CH4
+        PB6     ------> TIM4_CH1
+        */
+        GPIO_InitStruct.Pin = wheel_motor_back_left_pwm_Pin |
+                              wheel_motor_back_right_pwm_Pin |
+                              wheel_motor_front_right_pwm_Pin;
+        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
+        GPIO_InitStruct.Pull      = GPIO_NOPULL;
+        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
+        GPIO_InitStruct.Alternate = GPIO_AF2_TIM4;
+        HAL_GPIO_Init(GPIOD, &GPIO_InitStruct);
+
+        GPIO_InitStruct.Pin       = wheel_motor_front_left_pwm_Pin;
+        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
+        GPIO_InitStruct.Pull      = GPIO_NOPULL;
+        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
+        GPIO_InitStruct.Alternate = GPIO_AF2_TIM4;
+        HAL_GPIO_Init(wheel_motor_front_left_pwm_GPIO_Port, &GPIO_InitStruct);
+
+        /* USER CODE BEGIN TIM4_MspPostInit 1 */
+
+        /* USER CODE END TIM4_MspPostInit 1 */
+    }
+}
+
+void HAL_TIM_PWM_MspDeInit(TIM_HandleTypeDef* tim_pwmHandle)
+{
+    if (tim_pwmHandle->Instance == TIM4)
+    {
+        /* USER CODE BEGIN TIM4_MspDeInit 0 */
+
+        /* USER CODE END TIM4_MspDeInit 0 */
+        /* Peripheral clock disable */
+        __HAL_RCC_TIM4_CLK_DISABLE();
+        /* USER CODE BEGIN TIM4_MspDeInit 1 */
+
+        /* USER CODE END TIM4_MspDeInit 1 */
+    }
+}
+
+/* USER CODE BEGIN 1 */
+
+/* USER CODE END 1 */
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
