+++ src/firmware_new/boards/frankie_v1/tim.h
@@ -1,0 +1,61 @@
+/**
+ ******************************************************************************
+ * File Name          : TIM.h
+ * Description        : This file provides code for the configuration
+ *                      of the TIM instances.
+ ******************************************************************************
+ * @attention
+ *
+ * <h2><center>&copy; Copyright (c) 2020 STMicroelectronics.
+ * All rights reserved.</center></h2>
+ *
+ * This software component is licensed by ST under Ultimate Liberty license
+ * SLA0044, the "License"; You may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at:
+ *                             www.st.com/SLA0044
+ *
+ ******************************************************************************
+ */
+/* Define to prevent recursive inclusion -------------------------------------*/
+#ifndef __tim_H
+#define __tim_H
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+/* Includes ------------------------------------------------------------------*/
+#include "main.h"
+
+    /* USER CODE BEGIN Includes */
+
+    /* USER CODE END Includes */
+
+    extern TIM_HandleTypeDef htim4;
+
+    /* USER CODE BEGIN Private defines */
+
+    /* USER CODE END Private defines */
+
+    void MX_TIM4_Init(void);
+
+    void HAL_TIM_MspPostInit(TIM_HandleTypeDef *htim);
+
+    /* USER CODE BEGIN Prototypes */
+
+    /* USER CODE END Prototypes */
+
+#ifdef __cplusplus
+}
+#endif
+#endif /*__ tim_H */
+
+/**
+ * @}
+ */
+
+/**
+ * @}
+ */
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
