+++ src/firmware_new/boards/frankie_v1/usart.c
@@ -1,0 +1,117 @@
+/**
+ ******************************************************************************
+ * File Name          : USART.c
+ * Description        : This file provides code for the configuration
+ *                      of the USART instances.
+ ******************************************************************************
+ * @attention
+ *
+ * <h2><center>&copy; Copyright (c) 2020 STMicroelectronics.
+ * All rights reserved.</center></h2>
+ *
+ * This software component is licensed by ST under Ultimate Liberty license
+ * SLA0044, the "License"; You may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at:
+ *                             www.st.com/SLA0044
+ *
+ ******************************************************************************
+ */
+
+/* Includes ------------------------------------------------------------------*/
+#include "usart.h"
+
+/* USER CODE BEGIN 0 */
+
+/* USER CODE END 0 */
+
+UART_HandleTypeDef huart3;
+
+/* USART3 init function */
+
+void MX_USART3_UART_Init(void)
+{
+    huart3.Instance                    = USART3;
+    huart3.Init.BaudRate               = 115200;
+    huart3.Init.WordLength             = UART_WORDLENGTH_8B;
+    huart3.Init.StopBits               = UART_STOPBITS_1;
+    huart3.Init.Parity                 = UART_PARITY_NONE;
+    huart3.Init.Mode                   = UART_MODE_TX_RX;
+    huart3.Init.HwFlowCtl              = UART_HWCONTROL_NONE;
+    huart3.Init.OverSampling           = UART_OVERSAMPLING_16;
+    huart3.Init.OneBitSampling         = UART_ONE_BIT_SAMPLE_DISABLE;
+    huart3.Init.ClockPrescaler         = UART_PRESCALER_DIV1;
+    huart3.AdvancedInit.AdvFeatureInit = UART_ADVFEATURE_NO_INIT;
+    if (HAL_UART_Init(&huart3) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    if (HAL_UARTEx_SetTxFifoThreshold(&huart3, UART_TXFIFO_THRESHOLD_1_8) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    if (HAL_UARTEx_SetRxFifoThreshold(&huart3, UART_RXFIFO_THRESHOLD_1_8) != HAL_OK)
+    {
+        Error_Handler();
+    }
+    if (HAL_UARTEx_DisableFifoMode(&huart3) != HAL_OK)
+    {
+        Error_Handler();
+    }
+}
+
+void HAL_UART_MspInit(UART_HandleTypeDef* uartHandle)
+{
+    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    if (uartHandle->Instance == USART3)
+    {
+        /* USER CODE BEGIN USART3_MspInit 0 */
+
+        /* USER CODE END USART3_MspInit 0 */
+        /* USART3 clock enable */
+        __HAL_RCC_USART3_CLK_ENABLE();
+
+        __HAL_RCC_GPIOD_CLK_ENABLE();
+        /**USART3 GPIO Configuration
+        PD8     ------> USART3_TX
+        PD9     ------> USART3_RX
+        */
+        GPIO_InitStruct.Pin       = STLK_RX_Pin | STLK_TX_Pin;
+        GPIO_InitStruct.Mode      = GPIO_MODE_AF_PP;
+        GPIO_InitStruct.Pull      = GPIO_NOPULL;
+        GPIO_InitStruct.Speed     = GPIO_SPEED_FREQ_LOW;
+        GPIO_InitStruct.Alternate = GPIO_AF7_USART3;
+        HAL_GPIO_Init(GPIOD, &GPIO_InitStruct);
+
+        /* USER CODE BEGIN USART3_MspInit 1 */
+
+        /* USER CODE END USART3_MspInit 1 */
+    }
+}
+
+void HAL_UART_MspDeInit(UART_HandleTypeDef* uartHandle)
+{
+    if (uartHandle->Instance == USART3)
+    {
+        /* USER CODE BEGIN USART3_MspDeInit 0 */
+
+        /* USER CODE END USART3_MspDeInit 0 */
+        /* Peripheral clock disable */
+        __HAL_RCC_USART3_CLK_DISABLE();
+
+        /**USART3 GPIO Configuration
+        PD8     ------> USART3_TX
+        PD9     ------> USART3_RX
+        */
+        HAL_GPIO_DeInit(GPIOD, STLK_RX_Pin | STLK_TX_Pin);
+
+        /* USER CODE BEGIN USART3_MspDeInit 1 */
+
+        /* USER CODE END USART3_MspDeInit 1 */
+    }
+}
+
+/* USER CODE BEGIN 1 */
+
+/* USER CODE END 1 */
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
